requirements.txt
Django==5.2.4
djangorestframework==3.16.0
python-telegram-bot==22.3
Pillow==10.4.0
django-filter==25.1
python-dotenv==1.1.1
nest-asyncio==1.6.0
requests==2.31.0 